requirements.txt
streamlit
openai-whisper
pandas
nltk
yt-dlp